fix sidebar bug
