+++ .eslintrc
@@ -8,6 +8,9 @@
     "browser" : true,
     "node": true
   },
+  "ecmaFeatures": {
+    "arrowFunctions": true
+  },
   "settings": {
     "import/parser": "babel-eslint",
     "import/resolve": {
@@ -24,6 +27,7 @@
   "rules": {
     "react/prefer-stateless-function": 0, // 暂时关闭组件纯函数验证
     "react/no-multi-comp": 0,
+    "import/no-extraneous-dependencies": 0, // 关闭显示 react 不在依赖的问题
     "import/default": 0,
     "import/no-duplicates": 0,
     "import/named": 0,
@@ -34,6 +38,8 @@
     "comma-dangle": 0,  // not sure why airbnb turned this on. gross!
     "indent": [2, 2, {"SwitchCase": 1}],
     "no-console": 0,
-    "no-alert": 0
+    "no-alert": 0,
+    "no-script-url": 0,
+    "arrow-body-style": [1, "as-needed"]
   }
 }
