+++ package.json
@@ -83,15 +83,15 @@
     "babel-plugin-transform-runtime": "^6.9.0",
     "babel-polyfill": "^6.9.1",
     "babel-preset-es2015": "^6.9.0",
-    "babel-preset-react-optimize": "^1.0.1",
     "babel-preset-react": "^6.11.1",
+    "babel-preset-react-optimize": "^1.0.1",
     "babel-preset-stage-0": "^6.5.0",
     "babel-register": "^6.9.0",
     "babel-runtime": "^6.9.2",
     "better-npm-run": "0.0.9",
+    "classnames": "~2.2.5",
     "css-loader": "^0.23.1",
     "cssnano": "^3.7.3",
-    "classnames": "~2.2.5",
     "debug": "^2.2.0",
     "extract-text-webpack-plugin": "^1.0.1",
     "file-loader": "^0.9.0",
@@ -110,7 +110,7 @@
     "less-loader": "^2.2.3",
     "normalize.css": "^4.2.0",
     "postcss-loader": "^0.9.1",
-    "react": "^15.2.1",
+    "react": "^15.3.0",
     "react-dom": "^15.2.1",
     "react-redux": "^4.4.5",
     "react-router": "^2.5.2",
@@ -123,6 +123,7 @@
     "style-loader": "^0.13.1",
     "url-loader": "^0.5.7",
     "webpack": "^1.13.1",
+    "whatwg-fetch": "^1.0.0",
     "yargs": "^4.8.0"
   },
   "devDependencies": {
