+++ src/components/Header/package.json
@@ -2,5 +2,5 @@
   "name": "Header",
   "version": "0.0.0",
   "private": true,
-  "main": "./Header.js"
+  "main": "./Header.jsx"
 }
