+++ src/components/Icon/package.json
@@ -2,5 +2,5 @@
   "name": "Icon",
   "version": "0.0.0",
   "private": true,
-  "main": "./Icon.js"
+  "main": "./Icon.jsx"
 }
