+++ src/components/Sidebar/Items.js
@@ -1,0 +1,36 @@
+export const menus = [
+  {
+    id: 1,
+    name: 'adf',
+    icon: 'user',
+    children: [
+      {
+        id: 1,
+        name: 'children1',
+        icon: 'cog',
+      },
+      {
+        id: 2,
+        name: 'children2',
+        icon: 'cogs',
+      },
+      {
+        id: 3,
+        name: 'children3',
+        icon: 'cog',
+      },
+    ]
+  }, {
+    id: 2,
+    name: 'test',
+    icon: 'bank',
+    children: []
+  }, {
+    id: 3,
+    name: 'test1',
+    icon: 'car',
+    children: []
+  }
+];
+
+export default menus;
