--- src/components/Sidebar/SidebarItem.js
@@ -1,13 +1,0 @@
-import React, {Component} from 'react';
-
-class SidebarItem extends Component {
-    render() {
-        return (
-            <div>
-                
-            </div>
-        );
-    }
-}
-
-export default SidebarItem;
