+++ src/components/Sidebar/package.json
@@ -2,5 +2,5 @@
   "name": "Sidebar",
   "version": "0.0.0",
   "private": true,
-  "main": "./Sidebar.js"
+  "main": "./Sidebar.jsx"
 }
