+++ src/routes/CoreLayout/index.jsx
@@ -3,10 +3,11 @@ import Header from '../../components/Header';
 import Sidebar from '../../components/Sidebar';
 import '../../styles/app.scss';
 import './CoreLayout.scss';
+
 export const CoreLayout = ({ children }) => (
   <div>
     <Header />
-    <div className="clearfix"></div>
+    <div className="clearfix" />
     <div className="page-container">
       <div className="page-sidebar-warpper">
         <Sidebar />
